// lib/careers/types.ts
export interface PersonalInfoShape {
  firstName: string
  lastName: string
  streetAddress: string
  addressLine2: string
  city: string
  stateProvince: string
  zipCode: string
  email: string
  phone: string
}

export interface JobDetailsShape {
  applyingFor: string
  howDidYouHear: string
  liftCarry: string
  physicalAbilityGeneral: string
  physicalAbilityExplanation: string
  performDuties: string
  performDutiesExplanation: string
  hasDriversLicense: boolean
  otherAbility: string
  experienceWith: Record<string, boolean> // e.g. { houseCleaning: true, officeCleaning: false }
  otherExperienceText: string
}

export interface AvailabilityShape {
  workLocations: Record<string, boolean> // e.g. { Agawam: true, Springfield: false }
  otherLocationText: string
  generalAvailability: string
  hoursPerWeek: string
  daysAvailable: Record<string, boolean> // e.g. { Monday: true, Tuesday: false }
  shiftsAvailable: Record<string, boolean> // e.g. { Morning: true, Afternoon: false }
  otherShiftText: string
  weekendsHolidays: string // 'Yes' / 'No' / ''
}

export interface EducationShape {
  educationLevel: string
}

export interface WorkHistoryEntry {
  companyName: string
  streetAddress: string
  city: string
  zipCode: string
  fromDate: string
  toDate: string
  position: string
  supervisor: string
  reasonForLeaving: string
  mayContact: string // 'Yes' / 'No' / ''
}

export interface WorkHistoryShape {
  howManyEmployers: string
  entries: WorkHistoryEntry[]
}

export interface ReferenceEntry {
  name: string
  relationship: string
  phone: string
}
export interface ReferencesShape {
  entries: ReferenceEntry[]
}

export interface MilitaryServiceShape {
  branch: string
  rankAtDischarge: string
  serviceFrom: string
  serviceTo: string
  typeOfDischarge: string
}

export interface GettingToKnowYouShape {
  knowAnyone: string
  hardestJob: string
  magicSpell: string
  shorterWorkday: string
  winSomeoneOver: string
  addOneThingRestroom: string
  helpedCoworker: string
  ruleDisagree: string
  lastBossDescription: string
  fitInWell: string
  getOutOfBed: string
  busyOrDowntime: string
  // weekendsHolidays is in AvailabilityShape now
  oneHourConversation: string
  wishWeAsked: string
}

export interface StatementsShape {
  drugTestingAck: boolean
  applicantStatementAck: boolean
  partTimeStatementAck: boolean
}

export interface UploadsShape {
  driversLicense: File | null
  resume: File | null
  coverLetterText: string
}

export interface FormDataShape {
  personalInfo: PersonalInfoShape
  jobDetails: JobDetailsShape
  availability: AvailabilityShape
  education: EducationShape
  workHistory: WorkHistoryShape
  references: ReferencesShape
  militaryService: MilitaryServiceShape
  gettingToKnowYou: GettingToKnowYouShape
  statements: StatementsShape
  uploads: UploadsShape
}
// ...
export type SectionError = Record<string, string>